gcp/function/requirements.txt
functions-framework>=3.5.0
google-cloud-firestore>=2.14.0
flask>=2.3